Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -8,7 +8,7 @@ int main() {
     // printf("%s", welcome());
     int a,b,c;
     scanf("%d %d %d",&a,&b,&c);
-    int average = (a+b+c)/2;
+    float average = (a+b+c)/2;
     printf("Average: %.2f",average);
     return 0;
 }
